+++ Headers/Front/Interfaces/InterfacesManager.hpp
@@ -18,7 +18,7 @@ namespace Vox::Front::Interfaces
 			void RegisterInterface(const std::string &key, AInterface *inte);
 			void Render();
 
-			void HandleMouseMove(const size_t &xPos, const size_t &yPos) const;
+			void HandleMouseMove(const float &xPos, const float &yPos) const;
 			void HandleMouseClick(const int &button, const int &action) const;
 			void HandleMouseScroll(const double &xOff, const double &yOff) const;
 			InterfacesManager();
